LUCENE-7951:  Merge test in Geo3drptTest
